Refactor code for Pokemon validation and display
- Updated the validation logic in the POST route of the Pokemon validation API to properly filter and format invalid moves.
- Renamed the imported component in the Pokemon page to "PokemonTeamDisplay" for clarity.
- Added a new component "PokemonCard" to handle the display of individual Pokemon cards.
- Removed the deprecated "pokepaste" component.
