+++ frontend/app/pokemon/page.tsx
@@ -1,4 +1,4 @@
-import PokemonTeamDisplay from "@/components/pokemon/pokepaste";
+import PokemonTeamDisplay from "@/components/pokemon/pokemon-team-display";
 
 export default async function Pokemon() {
   // const teams = (await getPokemonTeams()).data;
